+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.1 (unreleased)
 
-- Retry on 404 responses
+- Add configurable list of HTTP error codes to retry on
 
 ## 0.1.0